// models/dto/imports/facturas/ImportRowFacturaResponse.ts
import { EstadoAsignacion } from "@/models/enums/EstadoAsignacion";
import { EstadoFactura } from "@/models/enums/EstadoFactura";
import { EstadoFila } from "@/models/enums/EstadoFila";
import { MedioPago } from "@/models/enums/MedioPago";

export type ImportRowFacturaResponse = {
  id: number;
  rowIndex: number;
  estado: EstadoFila;
  version: number;
  estadoAsignacion: EstadoAsignacion;
  asignadoAId: number | null;
  asignadoANombre: string | null;
  asignadoEn: string | null;
  actualizadoPorId: number | null;
  actualizadoEn: string | null;
  aplicadoPorId: number | null;
  aplicadoEn: string | null;
  errors: string | null;
  excelId: string | null;
  fechaCargaRaw: string | null;
  fechaEmisionRaw: string | null;
  fechaPagoRaw: string | null;
  fechaFacturaRaw: string | null;
  montoRaw: string | null;
  fechaCarga: string | null;
  fechaEmision: string | null;
  fechaPago: string | null;
  fechaFactura: string | null;
  comedorNombre: string | null;
  comedorId: number | null;
  numeroFactura: string | null;
  proveedorNombre: string | null;
  proveedorId: number | null;
  monto: number | null;
  numeroOperacion: string | null;
  bancoPagadorRaw: string | null;
  bancoNombre: string | null;
  bancoSociedadNombre: string | null;
  bancoId: number | null;
  formaPagoRaw: string | null;
  medioPago: MedioPago | null;
  estadoFacturaFinal: EstadoFactura | null;
  facturaIdAplicada: number | null;
};